+++ src/services/algorithms/index.ts
@@ -1,2 +1,3 @@
 export * from './generationAlgorithmsService';
 export * from './HlpPrimeGeneration';
+export * from './traversalAlgorithmsService';